moved editor code to editor folder
